+++ Sources/NablaCore/Errors/NetworkError.swift
@@ -1,0 +1,9 @@
+import Foundation
+
+public class NetworkError: NablaError {
+    public let message: String?
+    
+    public init(message: String? = nil) {
+        self.message = message
+    }
+}